refactor(frontend): add insights components for weekly spending, anomalies, income sources, and spending categories with empty state handling
