+++ frontend/src/components/insights/index.ts
@@ -1,0 +1,7 @@
+export { AnomaliesCard } from "@/components/insights/AnomaliesCard";
+export { BudgetPerformance } from "@/components/insights/BudgetPerformance";
+export { IncomeExpensesChart } from "@/components/insights/IncomeExpensesChart";
+export { IncomeSourcesChart } from "@/components/insights/IncomeSourcesChart";
+export { SpendingCategories } from "@/components/insights/SpendingCategories";
+export { SummaryCards } from "@/components/insights/SummaryCards";
+export { WeeklySpending } from "@/components/insights/WeeklySpending";
